feat: 공고 상세 Data hook으로 분리
